fix(registry): make GCR REGISTRY_LOCATION value consistent
The comments in `tpl/generate_params.toml` and the workflow registry
docs[1] state that the valid options for `registry_location` are:
`on-cluster`, `off-cluster`, `ecr`, `gcr`. This was correct when
`registry_location` is configured in `tpl/generate_params.toml`. When
defined by the `REGISTRY_LOCATION` enviroment variable, the value `gcs`
was accepted rather than `gcr`. This commit corrects this behavior so it
is consistent with both the documentation and the file-based definition.

[1] https://deis.com/docs/workflow/installing-workflow/configuring-registry/